requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
h5py>=3.7.0
opencv-python>=4.7.0
matplotlib>=3.5.0
tqdm>=4.64.0
PyYAML>=6.0
requests>=2.28.0